fix: Move useRBAC hook to separate file to resolve Fast refresh linting error
- Create client/src/hooks/useRBAC.js for the useRBAC hook
- Remove useRBAC export from RBACContext.jsx
- Update imports in Navbar.jsx and ProtectedRoute.jsx to use the new hook location
