+++ client/src/context/RBACContext.jsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useMemo } from "react";
+import React, { createContext, useMemo } from "react";
 import { hasPermission, getRolePermissions } from "../utils/rbacPermissions.js";
 
 const RBACContext = createContext(null);
@@ -15,10 +15,4 @@ export const RBACProvider = ({ children, userRole }) => {
   return <RBACContext.Provider value={value}>{children}</RBACContext.Provider>;
 };
 
-export const useRBAC = () => {
-  const context = useContext(RBACContext);
-  if (!context) {
-    throw new Error("useRBAC must be used within a RBACProvider");
-  }
-  return context;
-};
+export default RBACContext;
